Implement admin login client-side
Add isAdmin property to JWT. JWT is decoded by client and isAdmin is stored separately in Redux store. Also added AdminRoute component that functions just like a Route, but redirects users to the homepage if they're not admin. Added demo admin-only page protected by AdminRoute and hidden in the sidebar for volunteers. Also, slight refactor of actions and reducers.